Fifth part of the new transform system
It contains:

- Timestamp Replay and all the transforms that required for this replay type
- Refactor of the replay file as this is the last replay type that we have

This will be followed by a cleanup to remove the remnants of the old transform system.

For more information about the new transform system:
go/transform-refactor-first-stage

For Q&A:
go/agi-transform-review
