+++ gapis/api/vulkan/BUILD.bazel
@@ -81,9 +81,9 @@ go_library(
         "mem_binding_list.go",
         "memory_breakdown.go",
         "primeable_image_data.go",
-        "query_timestamps.go",
         "queue_task.go",
         "replay.go",
+        "replay_types.go",
         "resources.go",
         "scratch_resources.go",
         "state.go",
@@ -101,6 +101,7 @@ go_library(
         "transform_mapping_exporter.go",
         "transform_overdraw.go",
         "transform_profiling_layers.go",
+        "transform_query_timestamps.go",
         "transform_read_framebuffer.go",
         "transform_vulkan_terminator.go",
         "transform_wait_for_perfetto.go",
