+++ gapis/replay/BUILD.bazel
@@ -23,7 +23,6 @@ go_library(
         "context.go",
         "custom.go",
         "doc.go",
-        "end_of_replay.go",
         "events.go",
         "executor.go",
         "export_replay.go",
@@ -31,10 +30,8 @@ go_library(
         "id.go",
         "interfaces.go",
         "manager.go",
-        "mapping_exporter.go",
         "replay.go",
         "timestamps.go",
-        "wait_for_fence.go",
     ],
     embed = [":replay_go_proto"],
     importpath = "github.com/google/gapid/gapis/replay",
